ecere/File; BufferedFile: Initial large file support

+++ compiler/bootstrap/ecere/bootstrap/BuiltInContainer.c
@@ -89,9 +89,9 @@ struct __ecereNameSpace__ecere__com__IteratorPointer;
 struct __ecereNameSpace__ecere__com__SerialBuffer
 {
 unsigned char *  _buffer;
-unsigned int count;
-unsigned int _size;
-unsigned int pos;
+size_t count;
+size_t _size;
+size_t pos;
 } ecere_gcc_struct;
 
 extern void *  __ecereNameSpace__ecere__com__eSystem_New(unsigned int size);
