+++ compiler/bootstrap/ecere/bootstrap/instance.c
@@ -630,9 +630,9 @@ unsigned int circ;
 struct __ecereNameSpace__ecere__com__SerialBuffer
 {
 unsigned char *  _buffer;
-unsigned int count;
-unsigned int _size;
-unsigned int pos;
+size_t count;
+size_t _size;
+size_t pos;
 } ecere_gcc_struct;
 
 extern void *  memset(void *  area, int value, size_t count);
